rename defaultState to state
